requirements.txt
Django==1.6.5
Pillow==2.5.1
South==1.0
argparse==1.2.1
beautifulsoup4==4.3.2
django-autocomplete-light==1.4.14
django-taggit==0.12
djangorestframework==2.3.14
lxml==3.3.5
requests==2.3.0
wsgiref==0.1.2
